refactor: coffeeCount type 모두 Long으로 통일

+++ backend/src/main/java/com/woowacourse/moragora/dto/CoffeeStatResponse.java
@@ -10,15 +10,15 @@ public class CoffeeStatResponse {
 
     private final Long id;
     private final String nickname;
-    private final int coffeeCount;
+    private final Long coffeeCount;
 
-    public CoffeeStatResponse(final Long id, final String nickname, final int coffeeCount) {
+    public CoffeeStatResponse(final Long id, final String nickname, final Long coffeeCount) {
         this.id = id;
         this.nickname = nickname;
         this.coffeeCount = coffeeCount;
     }
 
-    public static CoffeeStatResponse of(final User user, final int coffeeCount) {
+    public static CoffeeStatResponse of(final User user, final Long coffeeCount) {
         return new CoffeeStatResponse(
                 user.getId(), user.getNickname(), coffeeCount
         );
